requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asyncpg==0.26.0
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.81.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.1
pypika-tortoise==0.1.6
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.2
websockets==10.3
